[issue-381] move datetime conversion from parser to src level
