+++ src/parser/json/package_parser.py
@@ -20,9 +20,10 @@ from src.model.spdx_none import SpdxNone
 from src.parser.error import SPDXParsingError
 from src.parser.json.actor_parser import ActorParser
 from src.parser.json.checksum_parser import ChecksumParser
-from src.parser.json.dict_parsing_functions import append_parsed_field_or_log_error, datetime_from_str, \
+from src.parser.json.dict_parsing_functions import append_parsed_field_or_log_error, \
     raise_parsing_error_if_logger_has_messages, json_str_to_enum_name, construct_or_raise_parsing_error, \
     parse_field_or_log_error, parse_field_or_no_assertion_or_none, parse_field_or_no_assertion
+from src.datetime_conversions import datetime_from_str
 from src.parser.json.license_expression_parser import LicenseExpressionParser
 from src.parser.logger import Logger
